+++ spec/repl/ast_reader_spec.rb
@@ -2,13 +2,19 @@ require 'spec_helper'
 
 include Electr
 
+class TestReadline
+  def self.readline(a, b)
+    print a
+    "2 3"
+  end
+end
+
 describe ASTReader do
 
   describe ".run" do
 
     it 'prompts' do
-      allow(STDIN).to receive(:gets).and_return("2 3\n")
-      reader = ASTReader.new
+      reader = ASTReader.new(TestReadline)
 
       expect { reader.run }.to output("E--ast> ").to_stdout
     end